updated change

+++ python_introduction/simple_interest.py
@@ -2,4 +2,4 @@ principal = 1000
 rate = 0.05
 time = 3
 print("principal = 1000, rate = 0.05, and time = 3")
-print("The simple interest is:",principal * rate * time)
+print("The simple interest is:",float(principal * rate * time))
